+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.2] - 2019-08-22
+### Fixed
+- Issue #6: filtering by non-existing property comparison throws exception. 
+
+## [0.4.1] - 2019-08-21
+### Fixed
+- Numeric properties in objects can be addressed correctly (bracket notation only).
+
 ## [0.4.0] - 2019-08-12
 ### Added
 - Processor is now able to delete and replace parts of JSON document by query.